% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_app.R
\name{create_api_file}
\alias{create_api_file}
\title{Create Plumber API File}
\usage{
create_api_file(dest, filename)
}
\arguments{
\item{dest}{Set as project directory. Default is getwd()}

\item{filename}{provide a name for the Plumber API file}
}
\description{
Create Plumber API File
}
\examples{
create_api_file("api", "route")
}
