% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_app.R
\name{create_app_file}
\alias{create_app_file}
\title{Create shiny app.R file.
Holds the logic for UI and Server files}
\usage{
create_app_file(dest = getwd())
}
\arguments{
\item{dest}{This is the working directory; default set as getwd()}
}
\description{
Create shiny app.R file.
Holds the logic for UI and Server files
}
\examples{
create_app_file()
}
