% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_app.R
\name{create_dockerfile}
\alias{create_dockerfile}
\title{Create a dockerfile template.}
\usage{
create_dockerfile(dest = getwd())
}
\arguments{
\item{dest}{This is the working directory; default set as getwd()}
}
\description{
Create a dockerfile template.
}
\examples{
create_dockerfile()
}
