% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_app.R
\name{create_entrypoint_file}
\alias{create_entrypoint_file}
\title{Create Plumber Entrypoint.R File}
\usage{
create_entrypoint_file(dest)
}
\arguments{
\item{dest}{}
}
\description{
Create Plumber Entrypoint.R File
}
