% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_app.R
\name{create_global_file}
\alias{create_global_file}
\title{Create a global.R file}
\usage{
create_global_file(dest = getwd())
}
\arguments{
\item{dest}{This is the working directory; default set as getwd()}
}
\description{
Create a global.R file
}
\examples{
create_global_file(getwd())
}
