% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_app.R
\name{create_plumber}
\alias{create_plumber}
\title{Create Plumber App}
\usage{
create_plumber(route_name)
}
\arguments{
\item{route_name}{}
}
\description{
Create Plumber App
}
\examples{
create_plumber(route_name = "main")
}
