% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_app.R
\name{create_r_module_file}
\alias{create_r_module_file}
\title{Create a shiny app default file.}
\usage{
create_r_module_file(dest = getwd(), file_name)
}
\arguments{
\item{dest}{This is the working directory; default set as getwd()}

\item{file_name}{Provide a file name}
}
\description{
Create a shiny app default file.
}
\examples{
create_r_module_file(dest = "R/", file_name = "home")
}
