% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_app.R
\name{create_run_app_file}
\alias{create_run_app_file}
\title{Create a shiny runApp.R file.}
\usage{
create_run_app_file(dest = getwd())
}
\arguments{
\item{dest}{This is the working directory; default set as getwd()}
}
\description{
Create a shiny runApp.R file.
}
\examples{
create_run_app_file()
}
