% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_app.R
\name{create_shiny_app}
\alias{create_shiny_app}
\title{Shiny App Directory and File creation.
Creates the files in the project directory.}
\usage{
create_shiny_app()
}
\description{
Shiny App Directory and File creation.
Creates the files in the project directory.
}
\examples{
create_shiny_app()
}
