% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_app.R
\name{dir_create_function}
\alias{dir_create_function}
\title{A function to create directory in project folder}
\usage{
dir_create_function(dir_name)
}
\arguments{
\item{dir_name}{Provide a directory name}
}
\description{
A function to create directory in project folder
}
\examples{
dir_create_function("R")
}
