% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_app.R
\name{file_create_function}
\alias{file_create_function}
\title{A function to create a file in project folder}
\usage{
file_create_function(file_name, dest = getwd())
}
\arguments{
\item{file_name}{Provide a file name}

\item{dest}{This is the working directory; default set as getwd()}
}
\description{
A function to create a file in project folder
}
\examples{
file_create_function("app.R", getwd())
}
