% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_app.R
\name{create_api_file}
\alias{create_api_file}
\title{Create Plumber API File}
\usage{
create_api_file(dest, filename)
}
\arguments{
\item{dest}{}

\item{filename}{}
}
\description{
Create Plumber API File
}
\examples{
create_api_file("api", "route")
}
