% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_app.R
\name{create_api_route_file}
\alias{create_api_route_file}
\title{Create Plumber API route file}
\usage{
create_api_route_file(dest, app_name)
}
\arguments{
\item{dest}{Set as project directory. Default is getwd()}

\item{app_name}{Create a name of the new api file.}
}
\description{
Create Plumber API route file
}
\examples{
create_api_route_file(dest = getwd(), app_name = "main")
}
